examples: validate markdown examples
To ensure that examples in the specification match up with the json
schemas, we propose a validation approach that will extract the fenced
code blocks from the markdown document and validate them. Examples are
marked with specification mediatypes which are then validated against
the respective json shema doc.

From here, we can propose that all examples be part of the
specification. Each act of clarification would be represented within the
specification and validated against the schema.

The tool can be run from the root repository directory with the
following command:

```console
$ go run examples.go < manifest.md
ok	 application/vnd.oci.image.manifest.list.v1+json
ok	 application/vnd.oci.image.manifest.v1+json
```

This changeset should be followeb by the following tasks:

1. Integrate with top-level Makefile.
2. Move tools (validate.go, examples.go) to `cmd/` directory.
3. Run validation as part of travis CI testsuite.

Note that there seems to be a bug in the schema validator when using
recursive references. We can investigate this after this PR is merged.

Signed-off-by: Stephen J Day <stephen.day@docker.com>