canonicalization: Add recommendations for canonicalization
Spun off from [1], now that we seem to have reached a consensus for
"SHOULD canonical JSON" there.  I've set this up so we have space to
add canonicalization recommendations for other formats, although the
only other basic type discussed in this repository is a gzipped
tarball and that's more than I want to bite off at the moment ;).

[1]: https://github.com/opencontainers/image-spec/issues/259
     Subject: manifest json fields order

Signed-off-by: W. Trevor King <wking@tremily.us>