manifest: Explicitly unspecified attributes for the initial layer directory
Point out that the layer's initial empty directory has unspecified
attributes.  Layer authors interested in the ownership and other
attributes of the unpacked root directory should explicitly set an
entry for it.

This commit originally focused on requiring the initial directory to
be empty before the layers are unpacked into it.  But most of that was
picked up in fed7241f (manifest: clarify the order of layers,
2016-09-29, #349).  I'd recommended "match" instead of "be" to allow
folks to apply via a union filesystem or whatever. As long as the
finished product is right, compliance testers and users should be
satisfied.

Signed-off-by: W. Trevor King <wking@tremily.us>