manifest-list: `platform` is effective optional
Per the 2017-01-25 call, stevvooe pointed out that in implementation
this field is optional and sane defaults are assumed.

This pairs nicely with the index model of providing a list of
descriptors, that may not be platform specific objects (i.e. an XML file
for AppStream data).

For media that pertains to an image, like AppStream metadata, may need
to accompany a manifest-list, but is not a descriptor that points to a
manifest (i.e. and XML file).

This clarifies the behavior on handling unknown `mediaType` descriptors.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>