README: Link to MeetBot logs for meeting minutes
Rob had been keeping the wiki up to date with links to the MeetBot
logs, but linking directly to the MeetBot logs gives us one less thing
to maintain.  The meetings which are only in the wiki are 2015-07-22,
2015-08-05, 2015-08-12, 2015-08-26, 2015-09-02, 2015-09-09,
2015-09-16, 2015-09-24, 2015-09-30.  The last is in MeetBot, but it
was our first MeetBot meeting, so the MeetBot logs are not great.
However, *all* of those meetings predate image-spec (founded
2016-03-24 [1]), so I haven't bothered to link the runtime wiki at
all.

[1]: https://groups.google.com/a/opencontainers.org/forum/#!topic/tob/KGyPpu8YfNk
     Subject: ACTION: Proposal for a new project: OCI Image Format Spec (v3)
     Date: Sun, 20 Mar 2016 01:41:27 +0000
     Message-ID: <CAD2oYtNiOgCgLgZk--4bfrdywTmwP=bqSxvkSJ-rv=PTEoaTag@mail.gmail.com>

Signed-off-by: W. Trevor King <wking@tremily.us>